Lighting Fixes + Level Cleanup
Changed how the corridors are lit up in the level. The mesh still has lights on both sides however the actual light source is in the middle of the corridor, this allows better lighting and should prevent a few issues with Unreal having too many light sources in an area. Also removed legacy materials that were imported with the corridors and ships.