Blueprinted Corridor Models
Blueprinted the corridor models so that they act as prefabs and can easily be modified at one central source and then applied to all instances. Also edited the grunge metal texture to remove rust as this is not realistic in a space environment. Rebuilt the prototyping map using the new blueprinted corridor sections.