Improved Straight Corridor Section
Improved the straight corridor section by reapplying materials and UV unwrapping them. A few graphical issues have been fixed such as visible seams around the lights, along with seams on the walls of the corridor. The materials on the corridor have been altered as well. Currently the floor and ceiling have a basic metallic texture in order to prevent seams between instances of the corridor.