Corridor 2 Corner Remodelled
Remodelled a section of the Corridor 2 Corner in order to work better with the textures in future. More loop cuts have been added to the corridor floor in order to prevent strange scaling issues when UV Unwrapping. Lights have not yet been added as the size needs to be resolved.