Corridor 2 Straight and Corner Seamless Flow
Redone the UVs on the straight and corner sections to allow the corridor to flow freely without any seams. Lights still need to be added to the Corner model and the rest of the model needs to be UV unwrapped to work best. The material indexes have also been reset in order to make assigning materials easier as they are consistent across corridor models.