Cross Junction Remodelled
The Cross Junction has been remodelled with the new lights. Trying to make the textures truly seamless is not possible, a break in the texture flow of some description will have to be added. This will most likely be in the form of a door or a grate in the floor to break the pattern.