Corridor 2 Door Design
Marked out the various panels for the door using the Pencil Tool in blender. This will be used in the blockout of these panels and ultimately to model the door. The blockout of the door frame has been completed, although extensions may be needed in order to hide the door panels when it is open. The door panels should not exceed the bounding box and "leak" into other corridors.