Corridor 2 Door Rigging
Rigged the Corridor 2 Door ready for animation. Also exported the file into Unreal to test the skeletal structure was maintained. Also recalculated the normals on the panels in order to fix invisible walls etc. The materials still need to be reworked and unwrapped as the textures are stretched. The material index also needs to be updated for the panels as they are currently using the first available index (which is the floor material).