Corridor2 Door Panel Scaling
Scaled the panels on the door panel to reduce the amount of texture flickering when the panels collide. The panels now slide inside each other at different thicknesses rather than all scaling inside the same thickness, this stops the overlapping textures and flickering. Also renamed the animations to be concise.