Corridor2 Door Functional In Engine
Created a blueprint to allow the corridor2 door to function when the player approaches, currently the collisions for the door are non-functional as it is a skeletal mesh. This may require the collision system to be switched the simple rather than complex. This will be better in the long run for performance.