Thickened Corridor 2 Straight Walls
Thickened the walls of the straight corridor section in order to prevent light going through the back faces. This is essential in order to prevent strange lighting leaks. This process will have to be done to all corridor pieces. Also set the new default UE4 Level to the prototyping level.