Space Corridor 2 Renamed + Reorganised
Renamed the space corridor assets to Corridor 1 as the previous corridor 1 was never developed beyond the blockout and the corridors will no longer be used in space, but rather in a planetary / base-building setting. The models folders have also had new sub folders for groups of objects. Also removed the ship models from the game as they were not used.