Remodelled Corridor1 Straight
Remodelled the Corridor1 Straight section to make it low-poly as the subdivision modifier will no longer be used as it was causing line-up issues after the exterior of the corridors were added to prevent light leaking. Also fixed a few issues when remodelling such as inconsistencies with the light border thickness.