Basic Environment Controller
Created a basic environment controller in C++. This controller is taken from an old project. Currently it features a time and date system, seasons, temperature and manipulating celestial settings in the skysphere. The celestial settings have not yet been bound to the skysphere so these are not visible yet but they are being generated. The weather is currently always set to sunny until an improved method is implemented.