Environment Controller Cleanup / Code Removal
Removed unnecessary parts of the environment code that were not needed for this project or needed to be reimplemented as the code was messy or non-functional. Some elements will be entirely redesigned such as the weather system whereas others such as temperature or seasons may be entirely removed as they may be out of the scope of the project. The sun intensity is now dependent on the time of day rather than the season and this makes the day/night cycle look more natural.