Networked Character Body Colours
Networked the ability for players to change their character colours. UI will be added for this to allow complete control over character colours. This will most likely be in the form of RGB sliders with a preview colour. Also added null checks to the materials in case they were not initialised correctly, this happens when a new player joins and the game tries to replicate the colours before the materials have been created on the remote client in the begin play function.