Functional Colour Selection via Character Editor
The colour selection element of the character editor is now complete with live colour changes synced across the network. The colour preview in the character editor has been removed as it was deemed unnecessary, considering that the character material updates and shows what the character willl look like with the colour applied. The sliders have been replaced with spinboxes which show the value of the slider while also acting like a slider. The apply button has also been removed as the changes are now live, rather than having to be applied. The server client also does not appear to load the colours from its character correctly and it defaults to 0. This will be fixed later.