Fixed Character Editor Colour Slider Default Value Bug
Fixed a bug where the character editor would not load the correct value for the body colour and would revert to the default colour value of 0, 0, 0, 1. This was because the body colour was being set to the slider defaults (0, 0, 0) as the value changed events were being trigerred before the values for the slider could be set to the actual body colour of the character. This has been fixed by disabling the value changed events until the widget has finished its construction.