Basic Nameplates
Created a basic widget for nameplates. These nameplates currently show the health and name of the actor they are on. They need to be improved by having the network code on them rather then on the actor itself to allow better implementation later on. They are networked and work with the character editor as they are currently only used on the player characters. Also removed the corridor test map as it had many issues with lighting.