Added First Person Mode
Added a first person mode to the game. This will be the primary perspective while the player can still go into third person. When in first person the animations glitch out, resulting in the player always playing the animation of running forward, despite not doing this movement. This will need to be resolved using an animation blueprint / blendspace.