Created Base Character Model + Skeleton
Created an outline for the character model. This is the starting point of creating characters. This basic model was created in order to create the skeleton required to animate other character models in future, the skeleton may have more bones added to it such as more details for the more fingers rather than just the fingers combined into one bone