Basic Building Rotation
Added a system for building rotation with the currently bound Q and E keys. This is currently snapped to 90 degree rotations but this can be changed, particularly for objects that may want to be placed at any angle, a snapping mode could be enabled / disabled with a key press in future. A building / object selection mode still needs to be implemented however.