Basic Building Selection
Basic Building Selection has been implemented. The selection is currently based on numbers that are pressed. This feature respawns the building preview with the new settings applied. This current solution is temporary and will hopefully be better in future. UI also needs to be created for selection rather than using numbers and building rotation should be saved between changing models and entering and exiting build mode for the quality of life of the player.