Improved Building Snapping
Improved the building snapping by changing the line trace to a box trace, allowing the player to be less accurate. This is better as it means the player does not have to spend ages trying to line up the snapping. The rotation system has also been reworked in preparation for the building snap rotation fix.