More Building Snapping Improvements
Added ability to toggle snapping mode for grid and building anchor based snapping. The preview build snapping system has been removed and will be reworked as it did not handle building rotations particularly well. The new system will make use of anchor points, rather than the old system of offsets. This allows better control as anchor points can be created anywhere on the mesh, the automatically generated ones currently are on the center of the face of the bounding box. The anchor points are currently set to display in the world using debug drawings.