Basic Building Destruction
Added the ability to destroy buildings in the world. This allows the player to modify their structures after they have been placed. This feature has also been replicated across the network to work in multiplayer. Also fixed a bug where the build anchor snap mode could not be toggled due to the function not being called when the input was setup.