Converted Building Blueprint to C++
Began converting the building system into C++, starting with the actual building class. This is the foundation of all buildings and this class provides a lot of backend, particularly for calculating the snapping points of buildings. Created a BP_Building asset for future reference for any graphical features that may want to be added or for extended functionality that wouldn't work in C++. As a result of the reparenting of the building BP from the corridors, the prototyping map has been wiped of all corridors and these will have to be rebuilt.