Converted Building Preview Blueprint to C++
Converted the building preview logic into C++ for better performance. This will most likely need to be improved and refactored. Also moved the first person camera from being initialised in blueprint to being initialised in C++ so that it could be accessed in the building preview class. Also fixed a bug where the root component was not being set correctly on buildings.