Fixed Building Placement Validity Issue
Fixed an issue where if a building was being snapped to an anchor point, the validity would not be retested once it had been placed. This meant that more buildings could be placed inside of the new building. This has been fixed by refreshing the building after every building placement. Also optimised the material setting functions by only running them when the value has changed.