Rebuilt Prototyping Map
Rebuilt a corridor system on the prototyping map using the ingame building system and then copied the corridor layout into the map. This actual use case of the system demonstrated a few flaws that should be looked into and potentially improved upon. Firstly, rotation needs to be reworked and incorporated into the anchor point system as rotating each piece can be painful and it will be better for the user if the corridors are rotated to the correct direction when going around a corner. Secondly corridors should not be placeable where the player can not see with the anchors, for example on the roof when the player is building from inside.