Corridor2 Straight Model Blockout
Created a basic alternate corridor model. This corridor tile set is is larger and has a more generic shape. The building folders have also been reorganised into their different tilesets because of this addition. This tileset will also be easy to apply to rooms because of the basic nature of its shape. This corridor piece will have to be edited in some parts in order to prevent light leaks when viewing at a distance.