Corridor2 Straight Piece Detailing
Added detailing to the corridor2 Straight piece frame. The corridor has had several details added to it, mainly the panel designs on the walll pieces and the connection to the floor from that wall piece. The light connector has also been redesigned. The pipes have been raised to allow more detailing to continue around the corridor frame.