Baked Corridor2 Straight Section For Texturing + MatID Texture
The Corridor2 Straight Section has had the colour ids baked into a texture. This allows the corridor section to be worked on in Quixel Mixer to create PBR materials and for them to be exported into Unreal Engine easily. Different approaches were tried to get this to work, hence why the texture preperations were reverted and unreverted.in preview commits.