Textured Corridor2 Straight
Textured the corridor2 straight section using the 3D painting mode in Quixel Mixer. This is not a great method for doing this as the textures are bound to that model and cannot be used. An alternative method will be used in future involving UV Unwrapping. Another issue with this method is that the textures leak and the setup is process is long for a not-great result. The new method should fix these issues, while allowing a more modular workflow.