Basic Landscape Heightmap Test
Created a basic landscape heightmap, mainly for testing out whether it is a useable option in the project. Realistically, if this approach is used, the terrain will have to be edited heavily to remove any noise and artifacting that occurs. Also getting the correct scale from the satellite imagery is quite difficult. A new site for getting heightmaps will be trialed later.