Removed Heightmap Textures and Terrain
Removed the heightmap textures and the terrain as they felt artifical due to the large amount of noise and terracing going on. It also did not offer a high amount of control. Hand sculpting the terrain is probably the approach that will be taken as generating terrain via heightmap or software does not allow control over small details such as roads, flat areas for buildings or any other considerations to incorporate into a game. A flat terrain of the maximum size (8x8 KM) has been created which will serve as a map; this map is large to allow for oceans and mountains at the boundaries of the map to prevent the player leaving the map.