Basic Landscape Shape
Created a basic flat landscape shape for the map to be used in the project. It features a large landmass that will eventually be bordered off with mountains to prevent players escaping. It also features a coast which has a large island to one side, which will most likely be the first developed region.