Basic Landscape Blockout
Made a basic blockout of the landscape. The landscape has been retated so the coast with the island is to the South instead of the North. This was done by exporting ther heightmap, rotating it and then importing the changed heightmap. Mountains have been added in the North with a river flowing along the base of them and into a river mouth with a small island in the middle of it, which flows into the southern sea.